refactored records
