+++ components/records/PatientView.js
