+++ pages/records/patient-record/index.js
@@ -7,7 +7,7 @@ import {
   Header,
   PatientView,
   PrescriptionsTable,
-} from '@/pages/records/_components';
+} from '@/components/records';
 import withAuth from '@/utils/auth';
 import Router from 'next/router';
 import { Button } from '@/components/TextComponents/';
