+++ pages/records/patient-vitals/index.js
@@ -8,7 +8,7 @@ import {
   VitalsTable,
   PrescriptionsTable,
   Header,
-} from '@/pages/records/_components';
+} from '@/components/records';
 import withAuth from '@/utils/auth';
 import toast from 'react-hot-toast';
 import axiosInstance from '@/pages/api/_axiosInstance';
